+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.4.0] - Unreleased
 ### Added
 - Settings screen is fully functional and allows to customize custom alphabet settings.
+- Language selection for letters alphabet
 
 ### Changed
 - Improved UI in Settings.